requirements.txt
antiorm==1.2.1
APScheduler==3.10.1
asgiref==3.6.0
async-generator==1.10
attrs==22.1.0
bcrypt==3.2.2
beautifulsoup4==4.11.1
bottle==0.12.23
bottle-websocket==0.2.9
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
db==0.1.1
db-sqlite3==0.0.1
distlib==0.3.6
Django==4.1.7
djangorestframework==3.14.0
dnspython==2.2.1
dominate==2.5.2
Eel==0.14.0
email-validator==1.2.1
escape==1.1
exceptiongroup==1.0.1
filelock==3.8.0
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==0.4.6
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.18.2
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.1
newsapi==0.1.1
numpy==1.22.3
outcome==1.2.0
packaging==21.3
pandas==1.4.2
platformdirs==2.5.4
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pyshorteners==1.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
selenium==4.6.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
spotipy==2.19.0
SQLAlchemy==1.4.37
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
twilio==7.9.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.12
virtualenv==20.16.7
visitor==0.1.3
webdriver-manager==3.8.4
Werkzeug==2.0.0
whichcraft==0.6.1
wsproto==1.2.0
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.5.1
